# sql: introduce Transaction ID Cache