--- a/pkg/sql/contention/cluster_settings.go
+++ b/pkg/sql/contention/cluster_settings.go
@@ -1,0 +1,20 @@
+// Copyright 2021 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package contention
+
+import "github.com/cockroachdb/cockroach/pkg/settings"
+
+// TxnIDCacheMaxSize limits the maximum byte size can be used by the TxnIDCache.
+var TxnIDCacheMaxSize = settings.RegisterByteSizeSetting(
+	`sql.contention.txn_id_cache.max_size`,
+	"the maximum byte size TxnID cache will use",
+	64*1024*1024, // 64 MB
+).WithPublic()