@Azhng 2021-11-30
Hmm, the initial implementation here looks like we have some pretty significant performance impacts for the avg latency.
We are seeing significant increase of latency across the board. 

| Branch          | tpmC    | %efficiency | Avg (ms) | p50 (ms) | p75 (ms) | p90 (ms) | p99 (ms) | pMax (ms) |
|-----------------|---------|-------------|----------|----------|----------|----------|----------|-----------|
| 38fab           | 31322.0 | 97.4%       |  140.6   |  28.3    | 62.9     |  79.7    | 1744.8   | 34359.7   |
| 38fab           | 31343.0 | 97.5%       |  122.2   |  31.5    | 71.3     |  88.1    | 167.8    | 28991.0   |
| 38fab           | 31354.6 | 97.5%       |  108.0   |  30.4    | 71.3     |  92.3    | 184.5    | 30064.8   |
| txn-id-cache    | 31357.2 | 97.5%       |  108.7   |  41.9    | 151.0    |  260.0   | 1208.0   | 14495.5   |
| txn-id-cache    | 31331.5 | 97.5%       |  143.4   |  71.3    | 218.1    |  318.8   | 1342.2   | 20401.1   |
| txn-id-cache    | 31262.0 | 97.2%       |  174.9   |  75.5    | 243.3    |  419.4   | 1744.8   | 32212.3   |

However, microbenchmark shows only very small delta of performance penalty

```
name                                           old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-24       239µs ± 5%     237µs ± 3%   ~     (p=0.436 n=10+10)
FlowSetup/vectorize=true/distribute=false-24      228µs ± 4%     231µs ± 3%   ~     (p=0.247 n=10+10)
FlowSetup/vectorize=false/distribute=true-24      233µs ± 4%     232µs ± 5%   ~     (p=0.684 n=10+10)
FlowSetup/vectorize=false/distribute=false-24     222µs ± 3%     225µs ± 6%   ~     (p=0.165 n=10+10)

name                                           old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-24      30.4kB ± 1%    30.5kB ± 2%   ~     (p=0.814 n=9+9)
FlowSetup/vectorize=true/distribute=false-24     28.6kB ± 0%    28.6kB ± 0%   ~     (p=0.645 n=8+8)
FlowSetup/vectorize=false/distribute=true-24     37.5kB ± 5%    37.0kB ± 0%   ~     (p=0.661 n=10+9)
FlowSetup/vectorize=false/distribute=false-24    35.2kB ± 0%    35.1kB ± 0%   ~     (p=0.460 n=9+10)

name                                           old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-24         254 ± 0%       256 ± 4%   ~     (p=0.240 n=8+10)
FlowSetup/vectorize=true/distribute=false-24        240 ± 0%       240 ± 0%   ~     (p=0.471 n=9+9)
FlowSetup/vectorize=false/distribute=true-24        250 ± 0%       250 ± 0%   ~     (p=0.747 n=8+9)
FlowSetup/vectorize=false/distribute=false-24       236 ± 0%       236 ± 0%   ~     (all equal)
```

I'm not quite sure how to interpret this results, cc @yuzefovich  thoughts here?