@Azhng 2021-12-09
Progress Update Dec 9

Spending a lot more time tuning the microbenchmark and it seems like we are now able to have a good idea of the performance impact through pprof's mutex profile. I think this is in line with what the TPCC benchmark showed us:

Before:

<img width="647" alt="Screen Shot 2021-12-09 at 15 50 26" src="https://user-images.githubusercontent.com/9267198/145488271-6b5872c5-13a0-4f88-a2f4-b4b86ae1c2a0.png">

After:

<img width="925" alt="Screen Shot 2021-12-09 at 15 50 39" src="https://user-images.githubusercontent.com/9267198/145488281-fea8517d-2272-4aa9-a7fe-53d22cc822a6.png">


However, I still haven't been able to find a way to use the microbenchmark to reveal the impact on end-to-end latency:

```
$ benchstat ~/tmp/baseline/bench ~/tmp/txn-id-vanilla/bench                                                                                            [15:42:42]
name                                     old time/op    new time/op    delta
ConcurrentSelect1/concurrentConn=64-24      746µs ± 9%     756µs ± 6%    ~     (p=0.579 n=10+10)
ConcurrentSelect1/concurrentConn=128-24    1.48ms ± 7%    1.48ms ± 4%    ~     (p=0.739 n=10+10)
ConcurrentSelect1/concurrentConn=256-24    3.17ms ± 2%    3.25ms ± 4%    ~     (p=0.101 n=8+10)
ConcurrentSelect1/concurrentConn=512-24    7.70ms ± 3%    8.10ms ± 4%  +5.30%  (p=0.000 n=9+9)

name                                     old us/op      new us/op      delta
ConcurrentSelect1/concurrentConn=64-24        710 ± 8%       722 ± 6%    ~     (p=0.255 n=10+10)
ConcurrentSelect1/concurrentConn=128-24     1.35k ± 7%     1.35k ± 4%    ~     (p=0.926 n=10+10)
ConcurrentSelect1/concurrentConn=256-24     2.72k ± 2%     2.74k ± 8%    ~     (p=0.740 n=7+10)
ConcurrentSelect1/concurrentConn=512-24     5.62k ± 8%     5.91k ±20%    ~     (p=0.050 n=9+9)

name                                     old alloc/op   new alloc/op   delta
ConcurrentSelect1/concurrentConn=64-24      898kB ± 0%     903kB ± 0%  +0.62%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=128-24    1.82MB ± 1%    1.83MB ± 0%  +0.79%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=256-24    3.75MB ± 0%    3.78MB ± 0%  +0.79%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=512-24    8.21MB ± 0%    8.31MB ± 1%  +1.16%  (p=0.000 n=9+10)

name                                     old allocs/op  new allocs/op  delta
ConcurrentSelect1/concurrentConn=64-24      9.56k ± 0%     9.75k ± 0%  +1.99%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=128-24     19.3k ± 0%     19.7k ± 0%  +2.11%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=256-24     39.6k ± 0%     40.4k ± 0%  +2.06%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=512-24     84.9k ± 0%     86.8k ± 0%  +2.25%  (p=0.000 n=9+10)
```