@Azhng 2021-12-14
Dec 14th update,

Turns out the easiest way to capture the performance degradation is to directly measure the overhead of `recordTransaction` using a histogram metrics, and using Golang's benchmarking framework to report that metrics.


Here's the benchmark result, the new measurement here is `overhead(ns/op)` and we can easily observe the skyrocketing overhead cost of transaction stats collection. 

```
$ benchstat ~/tmp/baseline/bench ~/tmp/txn-id-vanilla/bench                                                                                            [12:47:33]
name                                    old time/op          new time/op          delta
ConcurrentSelect1/concurrentConn=24-24           453µs ±11%           471µs ±11%       ~     (p=0.165 n=10+10)
ConcurrentSelect1/concurrentConn=48-24          1.15ms ±18%          1.27ms ±17%       ~     (p=0.052 n=10+10)
ConcurrentSelect1/concurrentConn=64-24          2.00ms ±12%          2.27ms ±31%    +13.57%  (p=0.029 n=10+10)

name                                    old ms/op            new ms/op            delta
ConcurrentSelect1/concurrentConn=24-24            0.28 ± 8%            0.30 ± 6%       ~     (p=0.165 n=10+10)
ConcurrentSelect1/concurrentConn=48-24            0.53 ± 4%            0.56 ± 3%     +5.99%  (p=0.000 n=10+9)
ConcurrentSelect1/concurrentConn=64-24            0.68 ± 4%            0.73 ± 4%     +8.00%  (p=0.000 n=9+10)

name                                    old overhead(ns/op)  new overhead(ns/op)  delta
ConcurrentSelect1/concurrentConn=24-24           2.68k ± 4%          13.60k ±11%   +407.90%  (p=0.000 n=8+8)
ConcurrentSelect1/concurrentConn=48-24           4.13k ±12%          39.19k ±18%   +848.56%  (p=0.000 n=10+8)
ConcurrentSelect1/concurrentConn=64-24           4.69k ± 8%          66.90k ±26%  +1326.04%  (p=0.000 n=8+10)

name                                    old alloc/op         new alloc/op         delta
ConcurrentSelect1/concurrentConn=24-24           369kB ± 1%           374kB ± 1%     +1.40%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=48-24           759kB ± 2%           774kB ± 2%     +2.00%  (p=0.002 n=10+10)
ConcurrentSelect1/concurrentConn=64-24          1.04MB ± 1%          1.08MB ± 4%     +3.45%  (p=0.000 n=10+10)

name                                    old allocs/op        new allocs/op        delta
ConcurrentSelect1/concurrentConn=24-24           3.67k ± 0%           3.75k ± 0%     +2.12%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=48-24           7.48k ± 1%           7.65k ± 1%     +2.28%  (p=0.000 n=10+10)
ConcurrentSelect1/concurrentConn=64-24           10.2k ± 1%           10.5k ± 1%     +2.85%  (p=0.000 n=10+9)
```

This is a good starting point for us to use microbenchmark to approximate the result of the TPCC result.